swingcake.c Add the challenge solution in C, with focus on reducing LOC
This provides an improvement to the already existing solution by reducing the number of lines of code, as stated in the README. However, it comes with a do NOT try at home warning, as such practice shouldn't be taken up. This code is just an exercise of trying to make the code shorter.